+++ src/components/BarChart/components/index.ts
@@ -1,3 +1,2 @@
 export {TooltipContent, TooltipContentProps} from './TooltipContent';
 export {AnnotationLine} from './AnnotationLine';
-export {Bar} from './Bar';